pyproject.toml
[project]
name = "sentiment-analyzer"
version = "0.1.0"
description = "한국어 감성 분석기 (NSMC + klue/roberta-base) — Planner-Generator-Evaluator 하네스 기반"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch>=2.4.0",
  "transformers>=4.46.0",
  "datasets>=3.0.0",
  "scikit-learn>=1.5.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "loguru>=0.7.0",
  "pyyaml>=6.0.1",
  "matplotlib>=3.9.0",
  "seaborn>=0.13.0",
  "evaluate>=0.4.3",
]

[project.optional-dependencies]
serve = ["gradio>=4.44.0"]
dev = ["pytest>=8.3.0", "ruff>=0.6.0", "mypy>=1.11.0"]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
